requirements.txt
pandas
openpyxl
undetected-chromedriver
selenium
python-dotenv
openai
windows-curses